Prototipos_Tkinter/agenda_telefonica/requirements.txt
psycopg2
python-dotenv
Faker
